backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
numpy>=1.24.0
tensorflow>=2.13.0
pydantic>=2.0.0
